requirements.txt
# Core Flask Dependencies
Flask==2.3.3
Flask-CORS==4.0.0
mysql-connector-python==8.1.0

# Data Science and ML Libraries
numpy==1.24.3
pandas==2.1.4
scikit-learn==1.3.2
tensorflow==2.15.0
keras==2.15.0
opencv-python==4.8.1.78
Pillow==10.1.0
joblib==1.3.2

# Additional ML Libraries
xgboost==2.0.2
lightgbm==4.1.0
nltk==3.8.1
spacy==3.7.2
textblob==0.17.1
wordcloud==1.9.2
imbalanced-learn==0.11.0

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.15.0

# Utilities
requests==2.31.0
python-dotenv==1.0.0
aiofiles==23.2.1

# Optional Advanced Features
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
torch==2.1.0
torchvision==0.16.0
transformers==4.35.2
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
optuna==3.4.0
mlflow==2.8.1
